requirements.txt
langchain==0.2.11
langchain-chroma==0.1.2
langchain-community==0.2.10
langchain-core==0.2.24
langchain-openai==0.1.19
openai==1.40.2
python-dotenv==1.0.0